Revert "blog posts scrolling update"

+++ app/src/main/java/com/freshlybuilt/enduserapp/models/PostsResponse.java
@@ -9,8 +9,8 @@ import java.util.List;
 public class PostsResponse {
  private String status;
  private float count;
- private int count_total;
- private int pages;
+ private float count_total;
+ private float pages;
  @SerializedName("posts")
  @Expose
  private List<Posts> posts = null;
@@ -28,11 +28,11 @@ public class PostsResponse {
   return count;
  }
 
- public int getCount_total() {
+ public float getCount_total() {
   return count_total;
  }
 
- public int getPages() {
+ public float getPages() {
   return pages;
  }
 
@@ -56,11 +56,11 @@ public class PostsResponse {
   this.count = count;
  }
 
- public void setCount_total(int count_total) {
+ public void setCount_total(float count_total) {
   this.count_total = count_total;
  }
 
- public void setPages(int pages) {
+ public void setPages(float pages) {
   this.pages = pages;
  }
